+++ src/components/Shared/SpectralSettings/SpectralSettingsComponent.tsx
@@ -14,6 +14,7 @@ export class SpectralSettingsComponent extends React.Component<{
     disable: boolean;
     disableChannelOption?: boolean;
     secondaryAxisCursorInfoVisible?: boolean;
+    label?: string;
 }> {
     render() {
         const frame = this.props.frame;
@@ -34,7 +35,7 @@ export class SpectralSettingsComponent extends React.Component<{
 
         return (
             <React.Fragment>
-                <FormGroup label={"Coordinate"} inline={true} disabled={disableCoordinateSetting}>
+                <FormGroup label={this.props.label ? this.props.label : "Coordinate"} inline={true} disabled={disableCoordinateSetting}>
                     <HTMLSelect
                         disabled={disableCoordinateSetting}
                         value={frame && frame.spectralCoordinate ? frame.spectralCoordinate : ""}
@@ -53,7 +54,7 @@ export class SpectralSettingsComponent extends React.Component<{
                     </FormGroup>
                 )}
 
-                <FormGroup label={"System"} inline={true} disabled={disableSystemSetting}>
+                <FormGroup label={this.props.label ? " " : "System"} inline={true} disabled={disableSystemSetting}>
                     <HTMLSelect
                         disabled={disableSystemSetting}
                         value={frame && frame.spectralSystem ? frame.spectralSystem : ""}