+++ src/stores/FileBrowserStore/FileBrowserStore.ts
@@ -80,11 +80,14 @@ export class FileBrowserStore {
     // Save image
     @observable saveFilename: string = "";
     @observable saveFileType: CARTA.FileType = CARTA.FileType.CASA;
-    @observable saveSpectralRange: string[] = ["0", "0"];
+    @observable saveSpectralStart: number;
+    @observable saveSpectralEnd: number;
+    @observable saveSpectralStride: number = 1;
     @observable saveStokesOption: number;
     @observable saveRegionId: number;
     @observable saveRestFreq: Freq = {value: 0, unit: FrequencyUnit.MHZ};
     @observable shouldDropDegenerateAxes: boolean;
+    @observable saveIsNativeValue: boolean = true;
 
     private extendedDelayHandle: any;
 
@@ -305,10 +308,18 @@ export class FileBrowserStore {
     @action initialSaveSpectralRange = () => {
         const activeFrame = AppStore.Instance.activeFrame;
         if (activeFrame && activeFrame.numChannels > 1 && activeFrame.isSpectralChannel) {
-            const min = Math.min(activeFrame.channelValueBounds.max, activeFrame.channelValueBounds.min);
-            const max = Math.max(activeFrame.channelValueBounds.max, activeFrame.channelValueBounds.min);
-            const delta = (max - min) / (activeFrame.numChannels - 1);
-            this.saveSpectralRange = [min.toString(), max.toString(), delta.toString()];
+            if (activeFrame.spectralSystemsSupported.length > 0) {
+                this.setSaveSpectralStart(Math.min(activeFrame.channelValueBounds.max, activeFrame.channelValueBounds.min));
+                this.setSaveSpectralEnd(Math.max(activeFrame.channelValueBounds.max, activeFrame.channelValueBounds.min));
+            } else {
+                if (this.saveIsNativeValue) {
+                    this.setSaveSpectralStart(Math.min(activeFrame.channelValueBounds.max, activeFrame.channelValueBounds.min));
+                    this.setSaveSpectralEnd(Math.max(activeFrame.channelValueBounds.max, activeFrame.channelValueBounds.min));
+                } else {
+                    this.setSaveSpectralStart(0);
+                    this.setSaveSpectralEnd(activeFrame.numChannels - 1);
+                }
+            }
         }
     };
 
@@ -567,12 +578,12 @@ export class FileBrowserStore {
         }
     };
 
-    @action setSaveSpectralRangeMin = (min: string) => {
-        this.saveSpectralRange[0] = min;
+    @action setSaveSpectralStart = (start: number) => {
+        this.saveSpectralStart = start;
     };
 
-    @action setSaveSpectralRangeMax = (max: string) => {
-        this.saveSpectralRange[1] = max;
+    @action setSaveSpectralEnd = (end: number) => {
+        this.saveSpectralEnd = end;
     };
 
     @action setSelectedFiles = (selection: ISelectedFile[]) => {
@@ -638,6 +649,11 @@ export class FileBrowserStore {
         if (restFreqStore) {
             this.setSaveRestFreq(restFreqStore.headerRestFreq);
         }
+    };
+
+    @action setSaveIsNativeValue = (val: boolean) => {
+        this.saveIsNativeValue = val;
+        this.initialSaveSpectralRange();
     };
 
     @computed get saveRestFreqInHz(): number {