+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+* Fixed incorrect range values in the save image dialogue with non-standard spectral axis images.([#2225](https://github.com/CARTAvis/carta-frontend/issues/2225)).
+
 ## [4.0.0]
 
 ### Added